Asynchronous movie export
Dropped FBMovieTools in favour of the new class
FBQuickTimeExporter, which initializes an empty
QuickTime movie (always on the main thread), and
can subsequently add chunks of images from the
reel (on any thread).
Together with the FBProgressViewController, this
allows FBF to display a progress bar and preview
image while exporting.