Remove assertion regarding skin count
My guess is that the crashes caused by this assertion are due to parallelism: One thread builds the pipeline, the other tries to fill it before it's ready. Sounds wonky, but I've decided to let the app ignore the mismatch; if it's due to parallelism, it will work itself out a few frames later.